Fix bug causing script to only work in repo root.
Due to an oversight, the previous code looked in a hard-coded path
for the .git/config file.  The entire remote look-up has now been
changed to a call to git, then parsing the verbose remote list.  A
possible alternative is to acquire the correct path with
`git rev-parse --git-dir`, and then parse the config file.  I opted
for the single call as it requires fewer libraries and is hopefully
less susceptible to change.  Unfortunately there appears to be
no plumbing access to directly enumerate remotes, though once their
names are known we can use `git config remote.<remotename>.url`,
but this would require multiple calls.  I prefer to avoid this for
simplicity.

So, for now, a simple user-mode listing and some string-splitting.